requirements.txt
azure-core==1.32.0
azure-cosmos==4.9.0
azure-storage-blob==12.24.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
Flask==3.1.0
Flask-Dance==7.1.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
# Slett hvis dette ikke brukes
# mysql==0.0.3 !
# mysql-connector==2.2.9
# mysql-connector-python==9.2.0
# mysqlclient==2.2.7
oauthlib==3.2.2
opencv-python==4.11.0.86
pillow==11.1.0
pycparser==2.22
pyodbc==5.2.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
URLObject==2.4.3
Werkzeug==3.1.3
folium==0.19.5
geopy==2.4.1
geographiclib==2.0